+++ src/main/java/poomasi/domain/product/dto/ProductResponse.java
@@ -24,4 +24,27 @@ public record ProductResponse(
                 .categoryId(product.getCategoryId())
                 .build();
     }
+
+    @Builder
+    public static record ProductResponse(
+            long id,
+            String name,
+            Long price,
+            int stock,
+            String description,
+            String imageUrl,
+            long categoryId
+    ) {
+        public static ProductResponse fromEntity(Product product) {
+            return ProductResponse.builder()
+                    .id(product.getId())
+                    .name(product.getName())
+                    .price(product.getPrice())
+                    .stock(product.getStock())
+                    .description(product.getDescription())
+                    .imageUrl(product.getImageUrl())
+                    .categoryId(product.getCategoryId())
+                    .build();
+        }
+    }
 }
