+++ src/main/java/poomasi/domain/product/service/ProductFarmerService.java
@@ -6,11 +6,14 @@ import org.springframework.transaction.annotation.Transactional;
 import poomasi.domain.member.service.MemberService;
 import poomasi.domain.product._category.service.CategoryService;
 import poomasi.domain.product.dto.ProductRegisterRequest;
+import poomasi.domain.product.dto.ProductResponse;
 import poomasi.domain.product.dto.UpdateProductQuantityRequest;
 import poomasi.domain.product.entity.Product;
 import poomasi.domain.product.repository.ProductRepository;
 import poomasi.global.error.BusinessError;
 import poomasi.global.error.BusinessException;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -50,5 +53,25 @@ public class ProductFarmerService {
     private Product getProductByProductId(Long productId) {
         return productRepository.findById(productId)
                 .orElseThrow(() -> new BusinessException(BusinessError.PRODUCT_NOT_FOUND));
+    }
+
+    @Service
+    @RequiredArgsConstructor
+    public static class ProductService {
+        private final ProductRepository productRepository;
+
+
+        public List<poomasi.domain.product.dto.ProductResponse.ProductResponse> getAllProducts() {
+            return productRepository.findAllByDeletedAtIsNull()
+                    .stream()
+                    .map(poomasi.domain.product.dto.ProductResponse.ProductResponse::fromEntity)
+                    .toList();
+        }
+
+        public poomasi.domain.product.dto.ProductResponse.ProductResponse getProductByProductId(Long productId) {
+            return productRepository.findByIdAndDeletedAtIsNull(productId)
+                    .map(ProductResponse.ProductResponse::fromEntity)
+                    .orElseThrow(() -> new BusinessException(BusinessError.PRODUCT_NOT_FOUND));
+        }
     }
 }
